initial ThinningAlgorithm created
added a new variable to tree, "dist", which is used by the thinning algorithm.
Also added setter for cut since it appeared to be missing

Created a new class "SortByDist" which is needed to run the thinningAlgorithm

modified class "Tree"'s ToString method: made it so you can now see the "cut" variable and it's value

Added "safeDistance" variable to forest. It is the distance deemed safe between trees. Hard set to 5 right now for testing. Need to implement a way for user to specify (perhaps from file input? GUI?)

Added "ThinningAlgorithm" method to Forest. It marks every tree to be cut or not based on the "safeDistance" variable. I stepped through it a couple of times to make sure it works how it should, and while I do want to test it a lot more, it seems correct from what I've seen.