initial commit
this is a overhaul to add radius and height to trees.

Declared the new height and radius variables in the Tree class, added respective getters and setters. Also added new variables to Tree.toString()

In Forest class:

Altered forest generation to include height and radius.

Altered forest export/import to accommodate new variables.

Massively changed distance method. Trees used to be just 3D points, now they are 3D shapes. XYZ values represent the base of the tree, height extends upwards, ex: (tree with xyz = 0, 5, 0 and height = 15 will have the top of the tree at xyz = 0, 20, 0.)
Now the distance method will not calculated distance between the tree's bases, but the shortest distance between the closest part of the trees. However, this method still treats trees like they have no width. To account for this, the radius of the trees is subtracted from this distance, giving the true distance between trees!